+++ NEWS.md
@@ -19,7 +19,7 @@
 ## new
 
 - `fillHoles` method for SpatRaster
-
+- `chunk` method to run a non-memory-safe SpatRaster function in chunks
 
 # version 1.8-80